pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = "my ecommerce website"
authors = ["Ankur Shrivastava"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.0"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
sqlalchemy = "^2.0.37"
bcrypt = "^4.2.1"
psycopg2-binary = "2.9.9"
alembic = "^1.14.1"
python-jose = "^3.4.0"
python-multipart = "^0.0.20"
black = "^25.1.0"
isort = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
